# perf: async fs calls in resolve and package handling