--- a/packages/vite/src/node/optimizer/index.ts
+++ b/packages/vite/src/node/optimizer/index.ts
@@ -851,7 +851,7 @@ export async function addManuallyIncludedOptimizeDeps(
     for (let i = 0; i < includes.length; i++) {
       const id = includes[i]
       if (glob.isDynamicPattern(id)) {
-        const globIds = expandGlobIds(id, config)
+        const globIds = await expandGlobIds(id, config)
         includes.splice(i, 1, ...globIds)
         i += globIds.length - 1
       }