--- a/packages/vite/src/node/optimizer/resolve.ts
+++ b/packages/vite/src/node/optimizer/resolve.ts
@@ -25,7 +25,7 @@ export function createOptimizeDepsIncludeResolver(
     // 'foo > bar > baz' => 'foo > bar' & 'baz'
     const nestedRoot = id.substring(0, lastArrowIndex).trim()
     const nestedPath = id.substring(lastArrowIndex + 1).trim()
-    const basedir = nestedResolveBasedir(
+    const basedir = await nestedResolveBasedir(
       nestedRoot,
       config.root,
       config.resolve.preserveSymlinks,
@@ -42,11 +42,14 @@ export function createOptimizeDepsIncludeResolver(
 /**
  * Expand the glob syntax in `optimizeDeps.include` to proper import paths
  */
-export function expandGlobIds(id: string, config: ResolvedConfig): string[] {
+export async function expandGlobIds(
+  id: string,
+  config: ResolvedConfig,
+): Promise<string[]> {
   const pkgName = getNpmPackageName(id)
   if (!pkgName) return []
 
-  const pkgData = resolvePackageData(
+  const pkgData = await resolvePackageData(
     pkgName,
     config.root,
     config.resolve.preserveSymlinks,
@@ -160,14 +163,15 @@ function getFirstExportStringValue(
 /**
  * Continuously resolve the basedir of packages separated by '>'
  */
-function nestedResolveBasedir(
+async function nestedResolveBasedir(
   id: string,
   basedir: string,
   preserveSymlinks = false,
 ) {
   const pkgs = id.split('>').map((pkg) => pkg.trim())
   for (const pkg of pkgs) {
-    basedir = resolvePackageData(pkg, basedir, preserveSymlinks)?.dir || basedir
+    basedir =
+      (await resolvePackageData(pkg, basedir, preserveSymlinks))?.dir || basedir
   }
   return basedir
 }