… packages/vite/LICENSE.md
… packages/vite/src/node/__tests__/build.spec.ts