### Description

`getRealPath` in resolve is taking ~8% of the workload in https://github.com/yyx990803/vite-dev-server-perf, and it is calling `fs.realPathSync.native`. After this PR, it takes ~2% by using `fsp.realPath` (equivalent to the `.native` sync version).

I tested the speed improvement using the test in the repo instead of the cpuprofile, but it isn't stable on my machine. I think that even if we now have a ton of promises being generated, the non-blocking `fs` calls should still make this change worth it. We need to test in some other benchmarks.

The PR also changes a `fs.readFileSync` in `package.ts` to be async, and that triggered a whole bunch of async/await changes in the codebase too.

---

### What is the purpose of this pull request? <!-- (put an "X" next to an item) -->

- [ ] Bug fix
- [ ] New Feature
- [ ] Documentation update
- [x] Other