@patak-dev 2023-12-02
I don't know why Windows isn't happy (at least in CI) when using `fsp.realPath`. Using `fs.realPathSync.native` works, so for now we are back to sync for windows.

This makes this PR a bit more controversial, because it will mean a small regression on Windows due to the extra promises being generated. But a lot of them are needed due to the `fsp.readFile` change, and I ended up reducing the number of created promises at [refactor: avoid creating promises in tryCleanFsResolve](https://github.com/vitejs/vite/pull/15211/commits/0faf06c0cb4f9abb5ada86f2b1eee59cf33d900e). So I think this may still be a good step.

If someone with more Windows knowledge finds a way to make `fsp.realPath` work, appreciated! But now we can also do that in another PR.