@userquin 2023-12-03
It works on first run (`pnpm test-serve`), second run fails with same CI errors:

![imagen](https://github.com/vitejs/vite/assets/6311119/d6207375-2663-4c4e-884f-2513a2a104eb)