@patak-dev 2024-10-28
Closing as this PR has a lot of conflicts now, we can review if we want to apply similar changes later on but this isn't seems to be something we need to do right now.